Move Icons out of CP namespace since they are used on the frontend (e.g. icon fieldtype)
