+++ src/Http/View/Composers/JavascriptComposer.php
@@ -7,8 +7,8 @@ use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 use Statamic\CommandPalette\Category;
 use Statamic\Facades\CommandPalette;
-use Statamic\Facades\CP\Icon;
 use Statamic\Facades\CP\Toast;
+use Statamic\Facades\Icon;
 use Statamic\Facades\Preference;
 use Statamic\Facades\Site;
 use Statamic\Facades\User;
